+++ src/dialogs/__init__.py
@@ -1,6 +1,5 @@
 """Dialog components for the TLTD application."""
-from .basket_selector import BasketSelectorDialog
 from .description_editor import DescriptionEditorDialog
 from .help_screen import HelpScreen
 
-__all__ = ["BasketSelectorDialog", "DescriptionEditorDialog", "HelpScreen"]
+__all__ = ["DescriptionEditorDialog", "HelpScreen"]